9392: added html table for accessibility. Also refactored to reduce size of file
